рефакторинг api
