+++ src/api/apiConfig.ts
@@ -1,0 +1,6 @@
+export const apiConfig = {
+  baseURL: "https://easydev.club/api/v1/",
+  headers: {
+    "Content-Type": "application/json",
+  },
+};
