+++ src/api/apiInstance.ts
@@ -1,39 +1,14 @@
 import axios from "axios";
 import { Token } from "../types/types";
 import { store } from "../state_manager/store";
-import { logoutUser, refreshThunk } from "../state_manager/authSlice";
+import { refreshThunk } from "../state_manager/authSlice";
+import { apiConfig } from "./apiConfig";
 
 export const apiInstance = axios.create({
-  baseURL: "https://easydev.club/api/v1/",
-  headers: {
-    "Content-Type": "application/json",
-  },
+  ...apiConfig,
 });
-
-export const refreshInstance = axios.create({
-  baseURL: "https://easydev.club/api/v1/",
-  headers: {
-    "Content-Type": "application/json",
-  },
-});
-
-export const refresh = async (refreshToken: string) => {
-  try {
-    const result = await refreshInstance.post<Token>("auth/refresh", {
-      refreshToken: refreshToken,
-    });
-    localStorage.setItem("refreshToken", result.data.refreshToken);
-    return result.data;
-  } catch (error) {
-    throw error;
-  }
-};
 
 apiInstance.interceptors.request.use(async (config) => {
-  if (!localStorage.getItem("refreshToken")) {
-    store.dispatch(logoutUser());
-  }
-
   config.headers["Authorization"] = `Bearer ${
     store.getState().auth.accessToken
   }`;
