+++ src/api/refreshInstance.ts
@@ -1,0 +1,6 @@
+import axios from "axios";
+import { apiConfig } from "./apiConfig";
+
+export const refreshInstance = axios.create({
+  ...apiConfig,
+});
